Make methods internal, move helper.cs
